+++ DocuSignSDK.framework/Headers/DSMEnvelopeDefinition.h
@@ -1,0 +1,6 @@
+#import <Foundation/Foundation.h>
+#import "DSMEnvelopeTemplate.h"
+
+@interface DSMEnvelopeDefinition : DSMEnvelopeTemplate
+
+@end