+++ DocuSignSDK.framework/Headers/DSMNameValue.h
@@ -1,0 +1,20 @@
+#import <Foundation/Foundation.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+/*!
+ @class DSMNameValue
+ */
+@interface DSMNameValue : NSObject
+/*!
+ * @brief TODO:Doc
+ */
+@property (nonatomic, copy) NSString *name;
+/*!
+ * @brief Specifies the value of the tab. [optional]
+ */
+@property (nonatomic, copy) NSString *value;
+
+@end
+
+NS_ASSUME_NONNULL_END